// client/src/interfaces/dropdown.ts
export interface DropdownItem {
  id: string;
  name: string;
  imageUrl?: string;
}

export interface DropdownProps {
  id: string;
  title?: string;
  data: DropdownItem[];
  position?: 'bottom-right' | 'bottom-left' | 'top-right' | 'top-left';
  hasImage?: boolean;
  style?: string;
  selectedId?: string;
  onSelect?: (id: string) => void;
}